prodsentinel-fake-services/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
fastapi==0.128.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
importlib-metadata==8.7.1
    # via opentelemetry-api
opentelemetry-api==1.39.1
    # via
    #   -r requirements.in
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.39.1
    # via -r requirements.in
opentelemetry-semantic-conventions==0.60b1
    # via opentelemetry-sdk
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via uvicorn
pyyaml==6.0.3
    # via uvicorn
starlette==0.50.0
    # via fastapi
structlog==25.5.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   fastapi
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uvicorn==0.40.0
    # via -r requirements.in
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
zipp==3.23.0
    # via importlib-metadata
